fix: address code scanning alerts and style issues from PR #19 review
- LinuxSpinner: cast to double before multiplication to prevent overflow
- MacOSPlatform_Dialogs: remove useless response assignments, use ternary
- ToolBar: remove empty foreach loop (platform handles propagation)
- Button: use LINQ Where for radio deselection filtering
- MacOSMenu: make CGPoint struct and fields readonly
- MacOSMenuItem: simplify if/else to ternary for menu item init and key code
- MacOSFontHandle: simplify if/else to ternary for font selection
- TreeItem: simplify if/else to ternary for platform item creation
- PlatformTestExamples: remove unused variable
- BrowserTests: use 'using' to ensure dispose on exception

+++ src/SWTSharp/Platform/Linux/LinuxSpinner.cs
@@ -149,7 +149,7 @@ internal class LinuxSpinner : LinuxWidget, IPlatformSpinner
             {
                 _increment = value;
                 // Page increment is typically 10x line increment
-                gtk_spin_button_set_increments(_spinButton, _increment, _increment * 10);
+                gtk_spin_button_set_increments(_spinButton, _increment, (double)_increment * 10);
             }
         }
     }
